Added libk, and changed up lots of things

+++ libc/string/memset.c
@@ -3,6 +3,6 @@
 void* memset(void* str, const int c, size_t num)
 {
     unsigned char* strdest = (unsigned char*) str;
-    for(int i = 0; i < num; i++) strdest[i] = c;
+    for(size_t i = 0; i < num; i++) strdest[i] = c;
     return strdest;
 }
